+++ src/globals.c
@@ -6,7 +6,7 @@
 #include <limits.h>
 #include <stdlib.h>
 
-const size_t WORD_SIZE = sizeof(void *) * CHAR_BIT;
+const int WORD_SIZE = sizeof(void *) * CHAR_BIT;
 
 const char *const Z64C_COPYRIGHT = "Copyright (c) 2018-2019 Filippo Costa et al.";
 const char *const Z64C_LICENSE = "MPL-2.0";
